[kernel] Source code cast cleanups

+++ elks/arch/i86/drivers/char/dircon.c
@@ -170,7 +170,7 @@ void Console_set_vc(unsigned int N)
 
     SetDisplayPage(Visible);
     PositionCursor(Visible);
-    Current_VCminor = (int) N;
+    Current_VCminor = N;
 }
 
 struct tty_ops dircon_ops = {
